simple_api/requirements.txt
authlib>=1.0.1
Flask==2.2.2
python-dotenv==0.20.0
python-jose==3.3.0
flask-cors==3.0.10
requests>=2.28.1
six==1.16.0